tsdb_ecg_project/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "ECG Processing Application"
authors = ["cryptonome <rcp@cryptonome.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
wfdb = "^4.1.2"
numpy = "^2.1.1"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
